requirements.txt
django
djangorestframework
psycopg2
django-allauth
qrcode
Pillow
django-extensions
django-grappelli
factory_boy